intelli-credit/requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
docling>=2.0.0
pydantic>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
duckduckgo-search>=6.0.0
reportlab>=4.0.0
python-docx>=1.1.0
matplotlib>=3.8.0
plotly>=5.20.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
streamlit-agraph>=0.0.45
